refactor(appstate): split 1807-line AppState into 7 cohort extensions
Following the same same-type extension pattern that took VoiceFlow's
AppState from 1098 → 493 lines (grapeot/voiceflow#40) and our own
earlier `+VoiceFlowKit` slice (#56), peel the rest of OpenCode's
AppState along responsibility lines.

Main file: 1807 → 717 lines (-60%). The seven new extension files
together carry 1151 lines, so net diff is roughly neutral — the gain
is structural, not byte count.

| Extension                  | Lines | Responsibility |
|---------------------------|-------|-----------------|
| `+Drafts.swift`           | 39    | per-session input draft + selected-model map persistence |
| `+Permissions.swift`      | 55    | respond/refresh permissions + questions |
| `+Models.swift`           | 61    | model + agent selection, canonical model ID migration |
| `+FileTree.swift`         | 77    | file tree / status / search / content load |
| `+Messages.swift`         | 210   | load messages / older / diff, sendMessage, optimistic + remove |
| `+SSE.swift`              | 351   | SSE connect/dispatch, polled status reconciliation, activity-text debouncing, missing-session recovery |
| `+Sessions.swift`         | 358   | configure / testConnection / load*/create/fork/delete/abort/title, bootstrap sync |

Demoted ~70 `private` members (vars + methods + statics) to default
(internal) so the extensions can reach them. They're still all
module-internal — no API surface change.

State fields stay on the main `@Observable` class (drafts, sessions
store, message store, etc.) so SwiftUI observation works exactly as
before. Only behavior moves.

Tests: TEST SUCCEEDED (full OpenCodeClientTests passes).
