+++ OpenCodeClient/OpenCodeClient/AppState+Sessions.swift
@@ -1,0 +1,358 @@
+import Foundation
+import os
+
+/// Session-list CRUD + connection bootstrap.
+///
+/// - Connection lifecycle: `configure` saves credentials, `testConnection`
+///   probes `/health` and kicks off SSE on success.
+/// - Session list: `loadSessions` / `loadMoreSessions` / `refreshSessions`
+///   pull a window of sessions; `upsertSession` / `selectSession` /
+///   `createSession` / `forkSession` / `deleteSession` are user actions.
+/// - Bootstrap: `bootstrapSyncCurrentSession` runs after SSE
+///   reconnect to make sure the currently selected session is still valid
+///   and its derived state (messages, permissions, status) is fresh.
+extension AppState {
+    func fetchSessions(limit: Int) async throws -> [Session] {
+        let directory = effectiveProjectDirectory
+        let loaded = try await apiClient.sessions(directory: directory, limit: limit)
+        let archivedCount = loaded.filter { $0.time.archived != nil }.count
+        Self.logger.debug("loadSessions: directory=\(directory ?? "nil", privacy: .public) limit=\(limit, privacy: .public) count=\(loaded.count, privacy: .public) archived=\(archivedCount, privacy: .public) ids=\(loaded.prefix(5).map(\.id).joined(separator: ","), privacy: .public)")
+        return loaded
+    }
+
+    func toggleSessionExpanded(_ sessionID: String) {
+        if expandedSessionIDs.contains(sessionID) {
+            expandedSessionIDs.remove(sessionID)
+        } else {
+            expandedSessionIDs.insert(sessionID)
+        }
+    }
+
+    func upsertSession(_ session: Session) {
+        let existingIndex = sessions.firstIndex(where: { $0.id == session.id })
+        sessions.removeAll { $0.id == session.id }
+
+        let targetIndex: Int
+        if let existingIndex {
+            targetIndex = min(existingIndex, sessions.count)
+        } else {
+            targetIndex = 0
+        }
+
+        sessions.insert(session, at: targetIndex)
+    }
+
+    func configure(serverURL: String, username: String? = nil, password: String? = nil) {
+        // Keep raw user input; security normalization happens at request time.
+        self.serverURL = serverURL
+        self.username = username ?? ""
+        self.password = password ?? ""
+    }
+
+    func testConnection() async {
+        connectionError = nil
+
+        let info = Self.serverURLInfo(serverURL)
+        guard info.isAllowed, let baseURL = info.normalized else {
+            isConnected = false
+            connectionError = info.warning ?? L10n.t(.errorInvalidBaseURL)
+            return
+        }
+
+        await apiClient.configure(baseURL: baseURL, username: username.isEmpty ? nil : username, password: password.isEmpty ? nil : password)
+        do {
+            let health = try await apiClient.health()
+            isConnected = health.healthy
+            serverVersion = health.version
+            if isConnected {
+                connectSSE()
+            }
+        } catch {
+            isConnected = false
+            connectionError = error.localizedDescription
+        }
+    }
+
+    func loadProjects() async {
+        guard isConnected else { return }
+        isLoadingProjects = true
+        do {
+            projects = try await apiClient.projects()
+            serverCurrentProjectWorktree = (try? await apiClient.projectCurrent())?.worktree
+        } catch {
+            Self.logger.warning("loadProjects failed: \(error.localizedDescription)")
+            projects = []
+        }
+        isLoadingProjects = false
+    }
+
+    func loadSessions() async {
+        guard isConnected else { return }
+        do {
+            let loaded = try await fetchSessions(limit: loadedSessionLimit)
+            sessions = loaded
+            hasMoreSessions = loaded.count >= loadedSessionLimit
+
+            // Only auto-select first session if there's no persisted selection at all
+            // This handles the case of fresh install or after all sessions are deleted
+            if currentSessionID == nil, let first = sessions.first {
+                currentSessionID = first.id
+                applySavedModelForCurrentSession()
+            }
+        } catch {
+            connectionError = error.localizedDescription
+        }
+    }
+
+    func loadMoreSessions() async {
+        guard isConnected else { return }
+        guard hasMoreSessions else { return }
+        guard !isLoadingMoreSessions else { return }
+
+        isLoadingMoreSessions = true
+        let nextLimit = Self.nextSessionFetchLimit(current: loadedSessionLimit)
+        defer { isLoadingMoreSessions = false }
+
+        do {
+            let loaded = try await fetchSessions(limit: nextLimit)
+            loadedSessionLimit = nextLimit
+            sessions = loaded
+            hasMoreSessions = loaded.count >= loadedSessionLimit
+
+            if currentSessionID == nil, let first = sessions.first {
+                currentSessionID = first.id
+                applySavedModelForCurrentSession()
+            }
+        } catch {
+            connectionError = error.localizedDescription
+        }
+    }
+
+    func loadAgents() async {
+        guard isConnected else { return }
+        isLoadingAgents = true
+        do {
+            let loaded = try await apiClient.agents()
+            agents = loaded
+            if selectedAgentIndex >= visibleAgents.count && !visibleAgents.isEmpty {
+                selectedAgentIndex = 0
+            }
+        } catch {
+            Self.logger.warning("loadAgents failed: \(error.localizedDescription)")
+        }
+        isLoadingAgents = false
+    }
+
+    func refreshSessions() async {
+        guard isConnected else { return }
+        await loadSessions()
+        await syncSessionStatusesFromPoll()
+    }
+
+    func selectSession(_ session: Session) {
+        guard currentSessionID != session.id else { return }
+
+        // Generate new loading ID to invalidate any in-flight tasks from previous session
+        let loadingID = UUID()
+        sessionLoadingID = loadingID
+
+        messageStore.resetStreaming()
+        messages = []
+        partsByMessage = [:]
+        currentSessionID = session.id
+        applySavedModelForCurrentSession()
+
+        Task { [weak self] in
+            guard let self else { return }
+            // Check if this task is still current before proceeding
+            guard self.sessionLoadingID == loadingID else { return }
+
+            await self.refreshSessions()
+            guard self.sessionLoadingID == loadingID else { return }
+
+            await self.loadMessages()
+            guard self.sessionLoadingID == loadingID else { return }
+
+            await self.refreshPendingPermissions()
+            guard self.sessionLoadingID == loadingID else { return }
+
+            await self.refreshPendingQuestions()
+            guard self.sessionLoadingID == loadingID else { return }
+
+            self.syncModelFromMessageHistory()
+            await self.loadSessionDiff()
+            guard self.sessionLoadingID == loadingID else { return }
+
+            await self.loadSessionTodos()
+            guard self.sessionLoadingID == loadingID else { return }
+        }
+    }
+
+    func isBusySession(_ status: SessionStatus?) -> Bool {
+        guard let type = status?.type else { return false }
+        return type == "busy" || type == "retry"
+    }
+
+    func loadSessionTodos() async {
+        guard let sessionID = currentSessionID else { return }
+        do {
+            let todos = try await apiClient.sessionTodos(sessionID: sessionID)
+            sessionTodos[sessionID] = todos
+        } catch {
+            if await recoverFromMissingCurrentSessionIfNeeded(error: error, requestedSessionID: sessionID) {
+                return
+            }
+            // keep previous value if any
+        }
+    }
+
+    func createSession() async {
+        guard isConnected else { return }
+
+        let loadingID = UUID()
+        sessionLoadingID = loadingID
+
+        do {
+            let session = try await apiClient.createSession(title: nil)
+            guard sessionLoadingID == loadingID else { return }
+
+            Self.logger.debug("createSession: created id=\(session.id, privacy: .public) directory=\(session.directory, privacy: .public) effectiveProjectDir=\(self.effectiveProjectDirectory ?? "nil", privacy: .public)")
+
+            upsertSession(session)
+            currentSessionID = session.id
+            if let m = selectedModel {
+                selectedModelIDBySessionID[session.id] = m.id
+                persistSelectedModelMap()
+            }
+            messageStore.resetStreaming()
+            messages = []
+            partsByMessage = [:]
+        } catch {
+            guard sessionLoadingID == loadingID else { return }
+            connectionError = error.localizedDescription
+        }
+    }
+
+    func forkSession(messageID: String?) async {
+        guard isConnected else { return }
+        guard let sessionID = currentSessionID else { return }
+
+        let loadingID = UUID()
+        sessionLoadingID = loadingID
+
+        do {
+            let forked = try await apiClient.forkSession(sessionID: sessionID, messageID: messageID)
+            guard sessionLoadingID == loadingID else { return }
+
+            Self.logger.debug("forkSession: created id=\(forked.id, privacy: .public) from=\(sessionID, privacy: .public) messageID=\(messageID ?? "nil", privacy: .public)")
+
+            upsertSession(forked)
+            currentSessionID = forked.id
+            messageStore.resetStreaming()
+            messages = []
+            partsByMessage = [:]
+            await loadMessages()
+            guard sessionLoadingID == loadingID else { return }
+            syncModelFromMessageHistory()
+        } catch {
+            guard sessionLoadingID == loadingID else { return }
+            connectionError = error.localizedDescription
+        }
+    }
+
+    func deleteSession(sessionID: String) async throws {
+        let previousCurrentSessionID = currentSessionID
+        try await apiClient.deleteSession(sessionID: sessionID)
+
+        sessions.removeAll { $0.id == sessionID }
+        clearSessionScopedCaches(sessionID: sessionID)
+
+        let nextSessionID = Self.nextSessionIDAfterDeleting(
+            deletedSessionID: sessionID,
+            currentSessionID: previousCurrentSessionID,
+            remainingSessions: sessions
+        )
+
+        guard previousCurrentSessionID == sessionID else {
+            currentSessionID = nextSessionID
+            return
+        }
+
+        clearCurrentSessionViewState()
+        if let nextSessionID {
+            currentSessionID = nextSessionID
+            applySavedModelForCurrentSession()
+            await loadMessages()
+            await refreshPendingPermissions()
+            await loadSessionDiff()
+            await loadSessionTodos()
+            syncModelFromMessageHistory()
+        } else {
+            currentSessionID = nil
+            pendingPermissions = []
+        }
+    }
+
+    func bootstrapSyncCurrentSession(reason: String) async {
+        guard currentSessionID != nil else { return }
+        let start = Date()
+
+        await validateAndRecoverCurrentSession()
+
+        await loadMessages()
+        await refreshPendingPermissions()
+        await refreshPendingQuestions()
+        await syncSessionStatusesFromPoll()
+        let elapsedMs = Int(Date().timeIntervalSince(start) * 1000)
+        Self.logger.debug("bootstrapSync reason=\(reason, privacy: .public) elapsedMs=\(elapsedMs, privacy: .public) messages=\(self.messages.count, privacy: .public) permissions=\(self.pendingPermissions.count, privacy: .public)")
+    }
+
+    /// Check whether the current session still exists on the server.
+    /// If the session was deleted (e.g. server restarted), reset currentSessionID
+    /// so the next loadSessions call auto-selects a valid session.
+    func validateAndRecoverCurrentSession() async {
+        guard let sid = currentSessionID else { return }
+        do {
+            _ = try await apiClient.messages(sessionID: sid, limit: 1)
+        } catch {
+            guard case APIError.httpError(let statusCode, _) = error, statusCode == 404 else { return }
+            Self.logger.debug("bootstrapSync: current session \(sid) not found on server, resetting")
+            currentSessionID = nil
+            await loadSessions()
+        }
+    }
+
+    func syncSessionStatusesFromPoll(markMissingBusyAsIdle: Bool = true) async {
+        guard isConnected else { return }
+        guard let statuses = try? await apiClient.sessionStatus() else { return }
+        mergePolledSessionStatuses(statuses, markMissingBusyAsIdle: markMissingBusyAsIdle)
+    }
+
+    func abortSession() async {
+        guard let sessionID = currentSessionID else { return }
+        do {
+            try await apiClient.abort(sessionID: sessionID)
+        } catch {
+            if await recoverFromMissingCurrentSessionIfNeeded(error: error, requestedSessionID: sessionID) {
+                return
+            }
+            connectionError = error.localizedDescription
+        }
+
+        await syncSessionStatusesFromPoll(markMissingBusyAsIdle: true)
+        await loadMessages()
+        await loadSessionDiff()
+    }
+
+    func updateSessionTitle(sessionID: String, title: String) async {
+        do {
+            _ = try await apiClient.updateSession(sessionID: sessionID, title: title)
+            await refreshSessions()
+        } catch {
+            if await recoverFromMissingCurrentSessionIfNeeded(error: error, requestedSessionID: sessionID) {
+                return
+            }
+            connectionError = error.localizedDescription
+        }
+    }
+}
